+++ wrappa/resources/__init__.py
@@ -1,2 +1,3 @@
 from .healthcheck import Healthcheck
 from .predict import Predict
+from .predictor import Predictor
